requirements.txt
streamlit
scikit-learn
pdfminer.six
docx2txt
